+++ config.js
@@ -1,0 +1,5 @@
+var config = {};
+
+config.mongoURL = process.env.MONGO_URL || 'mongodb://localhost:27017/Raumgleiter';
+
+module.exports = config;