+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [RaumgleiterEngine](https://github.com/RaumgleiterRSS/RaumgleiterEngine).
 
+## [Unreleased]
+### Added
+- Config file
+- Collection handling
+- Basic Aggregator
+- More tests
+
+### Changed
+- Rewrite Parser
+
 ## [0.0.1] - 2016-05-28
 ### Added
 - Basic Parser